--- src/main/java/AddCommand.java
@@ -1,14 +1,0 @@
-class AddCommand extends Command {
-    private final Task task;
-
-    public AddCommand(Task task) {
-        this.task = task;
-    }
-
-    @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws OracleException {
-        tasks.addTask(task);
-        storage.save(tasks.getTasks());
-        ui.showAddedTask(task, tasks.size());
-    }
-}
