--- src/main/java/Command.java
@@ -1,6 +1,0 @@
-abstract class Command {
-    abstract void execute(TaskList tasks, Ui ui, Storage storage) throws OracleException;
-    boolean isExit() {
-        return false;
-    }
-}
