--- src/main/java/ExitCommand.java
@@ -1,11 +1,0 @@
-class ExitCommand extends Command {
-    @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showGoodbye();
-    }
-
-    @Override
-    boolean isExit() {
-        return true;
-    }
-}
