--- src/main/java/ListCommand.java
@@ -1,6 +1,0 @@
-class ListCommand extends Command {
-    @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showTasks(tasks.getTasks());
-    }
-}
