--- src/main/java/OracleException.java
@@ -1,5 +1,0 @@
-class OracleException extends Exception {
-  public OracleException(String message) {
-    super(message);
-  }
-}
