--- src/main/java/TaskType.java
@@ -1,4 +1,0 @@
-enum TaskType {
-    TODO, DEADLINE, EVENT
-}
-
