+++ src/main/java/oracle/Oracle.java
@@ -1,3 +1,15 @@
+package oracle;
+
+import oracle.command.Command;
+
+import oracle.common.OracleException;
+import oracle.common.Parser;
+import oracle.common.Storage;
+import oracle.common.Ui;
+
+import oracle.task.TaskList;
+
+
 public class Oracle {
     private final Storage storage;
     private final TaskList tasks;
