+++ src/main/java/oracle/command/Command.java
@@ -1,0 +1,13 @@
+package oracle.command;
+
+import oracle.common.OracleException;
+import oracle.common.Storage;
+import oracle.common.Ui;
+import oracle.task.TaskList;
+
+public abstract class Command {
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws OracleException;
+    public boolean isExit() {
+        return false;
+    }
+}
