+++ src/main/java/oracle/command/DeleteCommand.java
@@ -1,0 +1,22 @@
+package oracle.command;
+
+import oracle.common.OracleException;
+import oracle.common.Storage;
+import oracle.common.Ui;
+import oracle.task.Task;
+import oracle.task.TaskList;
+
+public class DeleteCommand extends Command {
+    private final int index;
+
+    public DeleteCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws OracleException {
+        Task removedTask = tasks.deleteTask(index);
+        storage.save(tasks.getTasks());
+        ui.showDeletedTask(removedTask, tasks.size());
+    }
+}
