+++ src/main/java/oracle/command/ExitCommand.java
@@ -1,0 +1,17 @@
+package oracle.command;
+
+import oracle.common.Storage;
+import oracle.common.Ui;
+import oracle.task.TaskList;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showGoodbye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
