+++ src/main/java/oracle/command/UnmarkCommand.java
@@ -1,0 +1,24 @@
+package oracle.command;
+
+import oracle.common.OracleException;
+import oracle.common.Storage;
+import oracle.common.Ui;
+
+import oracle.task.Task;
+import oracle.task.TaskList;
+
+public class UnmarkCommand extends Command {
+    private final int index;
+
+    public UnmarkCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws OracleException {
+        Task task = tasks.getTask(index);
+        task.markUndone();
+        storage.save(tasks.getTasks());
+        ui.showUnmarkedTask(task);
+    }
+}
