+++ src/main/java/oracle/common/OracleException.java
@@ -1,0 +1,7 @@
+package oracle.common;
+
+public class OracleException extends Exception {
+  public OracleException(String message) {
+    super(message);
+  }
+}
