+++ src/main/java/oracle/common/Ui.java
@@ -1,3 +1,7 @@
+package oracle.common;
+
+import oracle.task.Task;
+
 import java.util.List;
 import java.util.Scanner;
 
@@ -10,7 +14,7 @@ public class Ui {
 
     public void showWelcome() {
         showLine();
-        System.out.println("    Hello! I'm Oracle");
+        System.out.println("    Hello! I'm oracle.Oracle");
         System.out.println("    What can I do for you?");
         showLine();
     }
