+++ src/main/java/oracle/task/TaskType.java
@@ -1,0 +1,6 @@
+package oracle.task;
+
+public enum TaskType {
+    TODO, DEADLINE, EVENT
+}
+
