+++ src/main/java/oracle/task/Todo.java
@@ -1,4 +1,6 @@
-class Todo extends Task {
+package oracle.task;
+
+public class Todo extends Task {
     public Todo(String description) {
         super(description, TaskType.TODO);
     }
