+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-    Hello! I'm Oracle
+    Hello! I'm oracle.Oracle
     What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
